requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
coverage==4.3.4
coveralls==1.1
cryptography==1.8.1
docopt==0.6.2
Fabric3==1.13.1.post1
Flask==0.12
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
multidict==2.1.4
packaging==16.8
paramiko==2.1.2
psycopg2==2.7.1
py==1.4.33
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.4.0
PyYAML==3.12
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.9
uWSGI==2.0.15
vcrpy==1.10.5
vcrpy-unittest==0.1.6
Werkzeug==0.12.1
wrapt==1.10.10
yarl==0.10.0
